Fix import from renamed incidents directory
