+++ src/argus/htmx/constants.py
@@ -1,5 +1,5 @@
 from argus.htmx.dateformat.constants import DATETIME_DEFAULT, DATETIME_FORMATS, DATETIME_CHOICES
-from argus.htmx.incidents.constants import DEFAULT_PAGE_SIZE, ALLOWED_PAGE_SIZES, PAGE_SIZE_CHOICES
+from argus.htmx.incident.constants import DEFAULT_PAGE_SIZE, ALLOWED_PAGE_SIZES, PAGE_SIZE_CHOICES
 from argus.htmx.themes.constants import THEME_CHOICES, THEME_NAMES, THEME_DEFAULT
