+++ src/argus/htmx/dateformat/views.py
@@ -8,7 +8,7 @@ from django_htmx.http import HttpResponseClientRefresh
 
 from argus.auth.utils import get_or_update_preference
 
-from argus.htmx.incidents.views import HtmxHttpRequest
+from argus.htmx.incident.views import HtmxHttpRequest
 from .constants import DATETIME_FORMATS
 
 LOG = logging.getLogger(__name__)
