+++ src/argus/htmx/themes/views.py
@@ -9,7 +9,7 @@ from django_htmx.http import HttpResponseClientRefresh
 from argus.auth.utils import get_or_update_preference
 
 from argus.htmx.constants import THEME_NAMES
-from argus.htmx.incidents.views import HtmxHttpRequest
+from argus.htmx.incident.views import HtmxHttpRequest
 
 LOG = logging.getLogger(__name__)
 THEMES_MODULE = "argus.htmx"
