+++ src/argus/htmx/urls.py
@@ -2,7 +2,7 @@ from django.contrib.auth import views as django_auth_views
 from django.urls import path, include
 
 from .auth import views as auth_views
-from .incidents.urls import urlpatterns as incident_urls
+from .incident.urls import urlpatterns as incident_urls
 from .timeslot.urls import urlpatterns as timeslot_urls
 from .notificationprofile.urls import urlpatterns as notificationprofile_urls
 from .destination.urls import urlpatterns as destination_urls
